+++ Products/PluggableAuthService/events.py
@@ -15,11 +15,16 @@ from Acquisition import aq_parent
 from zope.component import adapter
 from zope.component import subscribers
 from zope.interface import implements
+from zope.interface import implementer
 
 from Products.PluggableAuthService.interfaces.authservice import IBasicUser
 from Products.PluggableAuthService.interfaces.events \
     import ICredentialsUpdatedEvent
 from Products.PluggableAuthService.interfaces.events import IPASEvent
+from Products.PluggableAuthService.interfaces.events \
+    import IPrincipalAddedToGroupEvent
+from Products.PluggableAuthService.interfaces.events \
+    import IPrincipalRemovedFromGroupEvent
 from Products.PluggableAuthService.interfaces.events \
     import IPrincipalCreatedEvent
 from Products.PluggableAuthService.interfaces.events \
@@ -38,6 +43,22 @@ class PASEvent(object):
     def __init__(self, principal):
         self.principal = principal
         self.object = principal
+
+
+@implementer(IPrincipalAddedToGroupEvent)
+class PrincipalAddedToGroup(PASEvent):
+
+    def __init__(self, principal, group_id):
+        super(PrincipalAddedToGroup, self).__init__(principal)
+        self.group_id = group_id
+
+
+@implementer(IPrincipalRemovedFromGroupEvent)
+class PrincipalRemovedFromGroup(PASEvent):
+
+    def __init__(self, principal, group_id):
+        super(PrincipalRemovedFromGroup, self).__init__(principal)
+        self.group_id = group_id
 
 
 class PrincipalCreated(PASEvent):